code optimication
make the code more beautiful and readable
