+++ CoronaCounter/Controller/AppDelegate.swift
