+++ CoronaCounter/View/Assets.xcassets/AppIcon.appiconset/1024.png
