+++ CoronaCounter/View/Assets.xcassets/AppIcon.appiconset/128.png
