+++ CoronaCounter/View/Assets.xcassets/AppIcon.appiconset/16.png
