+++ CoronaCounter/View/Assets.xcassets/AppIcon.appiconset/256.png
