+++ CoronaCounter/View/Assets.xcassets/AppIcon.appiconset/32.png
