+++ CoronaCounter/View/Assets.xcassets/AppIcon.appiconset/512.png
