+++ CoronaCounter/View/Assets.xcassets/AppIcon.appiconset/64.png
