+++ CoronaCounter/View/Assets.xcassets/AppIcon.appiconset/Contents.json
